modified payments model

+++ db/migrate/20210313023145_create_payments.rb
@@ -1,7 +1,7 @@
 class CreatePayments < ActiveRecord::Migration[6.1]
   def change
     create_table :payments do |t|
-      t.decimal :amount, null: false
+      t.integer :amount, null: false
       t.references :booking, null: false, foreign_key: true
 
       t.timestamps
